+++ R/utils.R
@@ -51,15 +51,18 @@ set_ssl <- function(verify_ssl, ...) {
 #' appropriate messages based on the status.
 #'
 #' @param resp An HTTP response object from the httr2 package.
+#' @param verbose A logical value indicating whether to print detailed messages. Default is TRUE.
 #' @keywords internal
 #' @noRd
 #' @return This function does not return a value. It is used for its side effects.
-check_status_code <- function(resp) {
+check_status_code <- function(resp, verbose = TRUE) {
   status_code <- httr2::resp_status(resp)
   if (!status_code %in% c(200L, 202L)) {
     cli::cli_abort("Request failed with status code: {status_code}")
   } else {
-    cli::cli_alert_info("Request succeeded with status code: {status_code}")
+    if (isTRUE(verbose)) {
+      cli::cli_alert_info("Request succeeded with status code: {status_code}")
+    }
   }
 }
 
@@ -81,7 +84,9 @@ check_internet <- function(verbose = TRUE) {
     cli::cli_abort("It seems that you are not connected to internet!")
     out <- FALSE
   } else {
-    cli::cli_alert_info("Internet connection OK...")
+    if (isTRUE(verbose)) {
+      cli::cli_alert_info("Internet connection OK...")
+    }
     out <- TRUE
   }
 }