+++ NEWS.md
@@ -1,7 +1,8 @@
 
 # extractox 0.1.1
 
-* fix: #15 now `extr_iris` extracts the correct number of chemicals with no repetition.
+* Fixed `extr_iris` extracts the correct number of chemicals with no repetition (#15}.
+* Fixed `extr_comtox` works when a single `download_items` different than `DTXCID` is selected (#17).
 
 # extractox 0.1.0